removes one abstraction level of the parser
